OpenMW: File renames to try at establishing some kind of a standard
