+++ 01 OpenMW/Data Files/scripts/TamrielData/player_restrict_equipment.lua
